Add FakeStripeRepository
