+++ payments-core-testing/src/main/java/com/stripe/android/testing/AbsFakeStripeRepository.kt
@@ -425,7 +425,7 @@ abstract class AbsFakeStripeRepository : StripeRepository {
         paymentMethods: List<String>,
         amount: Int,
         currency: String,
-        country: String,
+        country: String?,
         locale: String,
         requestOptions: ApiRequest.Options
     ): Result<PaymentMethodMessage> {
